+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - [Marek Kurańda](https://github.com/mjkuranda): Landing page.
 - [Marek Kurańda](https://github.com/mjkuranda): Application buttons.
+- [Marek Kurańda](https://github.com/mjkuranda): Initial `/search` page.
+
+### Removed
+- [Marek Kurańda](https://github.com/mjkuranda): `/explore` page.
 
 ## [1.0.0] - 2024-09-14
 ### Added